+++ packages/components/src/components/Action/index.ts
@@ -1,3 +1,4 @@
 export { Action } from "./Action";
 export * from "./types";
 export { default } from "./Action";
+export { useAriaAnnounceSuspense } from "./lib/ariaLive";