+++ packages/components/src/components/Checkbox/Checkbox.tsx
@@ -9,32 +9,49 @@ import clsx from "clsx";
 import type { PropsWithChildren } from "react";
 import * as Aria from "react-aria-components";
 import styles from "./Checkbox.module.scss";
+import { useFieldComponent } from "@/lib/hooks/useFieldComponent";
+import { PropsContextProvider } from "@/lib/propsContext";
+import { useObjectRef } from "@react-aria/utils";
+import { useMakeFocusable } from "@/lib/hooks/dom/useMakeFocusable";
 
 export interface CheckboxProps
   extends PropsWithChildren<Omit<Aria.CheckboxProps, "children">>,
-    FlowComponentProps<HTMLLabelElement> {}
+    FlowComponentProps {
+  inputClassName?: string;
+}
 
 /** @flr-generate all */
 export const Checkbox = flowComponent("Checkbox", (props) => {
-  const { children, className, ref, ...rest } = props;
+  const { children, className, ref, inputClassName, ...rest } = props;
 
-  const rootClassName = clsx(styles.checkbox, className);
+  const { FieldErrorView, fieldPropsContext, fieldProps } =
+    useFieldComponent(props);
+
+  const localCheckboxRef = useObjectRef(ref);
+  useMakeFocusable(localCheckboxRef);
 
   return (
-    <Aria.Checkbox {...rest} className={rootClassName} ref={ref}>
-      {({ isSelected, isIndeterminate }) => (
-        <>
-          {isSelected ? (
-            <IconCheckboxChecked className={styles.icon} />
-          ) : isIndeterminate ? (
-            <IconCheckboxIndeterminate className={styles.icon} />
-          ) : (
-            <IconCheckboxEmpty className={styles.icon} />
-          )}
-          {children}
-        </>
-      )}
-    </Aria.Checkbox>
+    <div
+      {...fieldProps}
+      className={clsx(styles.checkbox, className, fieldProps.className)}
+      ref={localCheckboxRef}
+    >
+      <Aria.Checkbox {...rest} className={clsx(inputClassName, styles.input)}>
+        {({ isSelected, isIndeterminate }) => (
+          <PropsContextProvider props={fieldPropsContext}>
+            {isSelected ? (
+              <IconCheckboxChecked className={styles.icon} />
+            ) : isIndeterminate ? (
+              <IconCheckboxIndeterminate className={styles.icon} />
+            ) : (
+              <IconCheckboxEmpty className={styles.icon} />
+            )}
+            {children}
+          </PropsContextProvider>
+        )}
+      </Aria.Checkbox>
+      <FieldErrorView />
+    </div>
   );
 });