+++ packages/components/src/components/MarkdownEditor/stories/Default.stories.tsx
@@ -40,8 +40,24 @@ export const WithFieldError: Story = {
   ),
 };
 
-export const Resizeable: Story = {
+export const AutoResizeable: Story = {
   args: { rows: 1, autoResizeMaxRows: 5 },
+};
+
+export const Resizeable: Story = {
+  args: { allowResize: true },
+};
+
+export const HorizontallyResizeable: Story = {
+  args: { allowHorizontalResize: true },
+};
+
+export const VerticallyResizeable: Story = {
+  args: { allowVerticalResize: true },
+};
+
+export const VerticallyAndAutoResizeable: Story = {
+  args: { allowVerticalResize: true, rows: 1, autoResizeMaxRows: 5 },
 };
 
 export const WithOnChange: Story = {