+++ packages/components/src/components/Overlay/components/index.ts
@@ -1,2 +1,3 @@
 import "./view";
 export { OverlayContent, type OverlayContentProps } from "./OverlayContent";
+export * from "./OverlaySuspenseFallback";