+++ packages/components/src/components/TextField/TextField.tsx
@@ -8,9 +8,11 @@ import type { FlowComponentProps } from "@/lib/componentFactory/flowComponent";
 import { flowComponent } from "@/lib/componentFactory/flowComponent";
 import type { PropsWithClassName } from "@/lib/types/props";
 import { ReactAriaControlledValueFix } from "@/lib/react/ReactAriaControlledValueFix";
+import { useFieldComponent } from "@/lib/hooks/useFieldComponent";
+import { PropsContextProvider } from "@/lib/propsContext";
 
 export interface TextFieldProps
-  extends Omit<TextFieldBaseProps, "input" | "className">,
+  extends Omit<TextFieldBaseProps, "FieldErrorView" | "input" | "className">,
     Pick<Aria.InputProps, "placeholder" | "form">,
     PropsWithClassName,
     FlowComponentProps<HTMLInputElement> {}
@@ -30,9 +32,19 @@ export const TextField = flowComponent("TextField", (props) => {
     </ReactAriaControlledValueFix>
   );
 
+  const { FieldErrorView, fieldPropsContext, fieldProps } =
+    useFieldComponent(props);
+
   return (
-    <TextFieldBase {...rest} input={input}>
-      {children}
+    <TextFieldBase
+      {...rest}
+      {...fieldProps}
+      FieldErrorView={FieldErrorView}
+      input={input}
+    >
+      <PropsContextProvider props={fieldPropsContext}>
+        {children}
+      </PropsContextProvider>
     </TextFieldBase>
   );
 });