+++ packages/components/src/integrations/react-hook-form/components/Field/stories/Switch.stories.tsx
@@ -1,14 +1,14 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import React from "react";
+import React, { useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { action } from "storybook/actions";
-import { Label } from "@/components/Label";
 import { Field, Form, typedField } from "@/integrations/react-hook-form";
 import { Button } from "@/components/Button";
 import { Section } from "@/components/Section";
 import { ActionGroup } from "@/components/ActionGroup";
 import { sleep } from "@/lib/promises/sleep";
 import { Switch } from "@/components/Switch";
+import { FieldError } from "@/components/FieldError";
 
 const submitAction = action("submit");
 
@@ -37,9 +37,7 @@ const meta: Meta<typeof Field> = {
       <Form form={form} onSubmit={handleOnSubmit}>
         <Section>
           <Field name="isEnabled">
-            <Switch>
-              <Label>Text</Label>
-            </Switch>
+            <Switch>Text</Switch>
           </Field>
 
           <ActionGroup>
@@ -56,21 +54,43 @@ type Story = StoryObj<typeof Field>;
 
 export const Default: Story = {};
 
-export const WithFocusAndError: Story = {
-  render: () => {
+export const WithFieldError: Story = {
+  render: (props) => {
+    const form = useForm();
+    useEffect(() => {
+      form.setError("field", {
+        type: "required",
+        message: "ErrorFromForm",
+      });
+    }, []);
+
+    return (
+      <Form form={form} onSubmit={async () => await sleep(2000)}>
+        <Field name={"field"}>
+          <Switch {...props}>Field1</Switch>
+        </Field>
+        <Switch>
+          Field2
+          <FieldError>ErrorFromOuterFieldError!</FieldError>
+        </Switch>
+      </Form>
+    );
+  },
+};
+
+export const WithFocus: Story = {
+  render: (props) => {
     const form = useForm();
     return (
       <Form form={form} onSubmit={async () => await sleep(2000)}>
-        <Field name={"email"}>
-          <Switch>
-            <Label>Text</Label>
-          </Switch>
+        <Field name={"field"}>
+          <Switch {...props}>Field</Switch>
         </Field>
         <div style={{ marginBottom: "2200px" }} />
         <Button
           onPress={() =>
             form.setError(
-              "email",
+              "field",
               { type: "required", message: "oh no" },
               { shouldFocus: true },
             )
@@ -78,7 +98,7 @@ export const WithFocusAndError: Story = {
         >
           err through form
         </Button>
-        <Button onPress={() => form.setFocus("email")}>
+        <Button onPress={() => form.setFocus("field")}>
           focus through form
         </Button>
         <Button type="submit">Submit</Button>