+++ packages/remote-elements/src/auto-generated/RemoteSliderElement.ts
@@ -24,6 +24,7 @@ export class RemoteSliderElement extends FlowRemoteElement<RemoteSliderElementPr
       inert: {},
       inputRef: {},
       isDisabled: {},
+      isInvalid: {},
       isReadOnly: {},
       lang: {},
       maxValue: {},